Eliminate HardwareConfigBase and improve Bus by using enum ordinal, sounds bad, but it really good.
