+++ src/main/java/frc/team555/hardware/DeviceEnum.java
@@ -1,0 +1,5 @@
+package frc.team555.hardware;
+
+public interface DeviceEnum {
+    int ordinal();
+}
