+++ src/main/java/frc/team555/hardware/PDBSlots.java
@@ -1,0 +1,9 @@
+package frc.team555.hardware;
+
+public class PDBSlots extends Bus<PDBSlots> {
+    public PDBSlots() {
+        super(1, 12);
+    }
+}
+
+
