+++ src/main/java/frc/team555/hardware/SRX.java
@@ -1,12 +1,10 @@
 package frc.team555.hardware;
 
 import com.ctre.CANTalon;
-import frc.team555.hardware.HardwareConfigBase.*;
-
 
 public class SRX extends CANTalon {
 
-    public SRX(PortID<HardwareConfigBase.CANBus> c) {
+    public SRX(PortID<CANBus> c) {
         super(c.get());
     }
 }
