+++ src/main/java/frc/team555/robot/HardwareConfig.java
@@ -1,15 +1,27 @@
 package frc.team555.robot;
 
 import com.ctre.CANTalon;
-import edu.wpi.first.wpilibj.PowerDistributionPanel;
 import frc.team555.hardware.*;
 
-public class HardwareConfig extends HardwareConfigBase {
+public class HardwareConfig {
 
+    // Build control system
+    public CANBus canBus = new CANBus();
+    public PDBSlots pdbSlots = new PDBSlots();
+
+    public PDB pdb = new PDB(pdbSlots);
+
+    // Create builders for types of devices we'll use
+    public SRXBuilder srxBuilder = new SRXBuilder(canBus);
+
+    // Build attached devices
+
+    // Create list of devices so that multiple uses of any device will be standardized
+    // For example, the FrontLeftDriveMotor is used in canBus, pdbSlots, and when creating the speed controller
+    // all of these are "immune" from spelling errors by using the enum.
     public enum Devices implements DeviceEnum {
         FrontLeftDriveMotor, FrontRightDriveMotor, BackLeftDriveMotor,BackRightDriveMotor;
     }
-
 
     public HardwareConfig() {
         canBus
@@ -26,6 +38,8 @@ public class HardwareConfig extends HardwareConfigBase {
     }
 
 
+    // These should probably be built like the devices above, but using
+    // a collection (not Bus) that is geared to hold another type (i.e. SpeedController)
     public CANTalon drivetrainFL = srxBuilder.build(Devices.FrontLeftDriveMotor);
     public CANTalon drivetrainFR = srxBuilder.build(Devices.FrontRightDriveMotor);
     public CANTalon drivetrainBL = srxBuilder.build(Devices.BackLeftDriveMotor);
